+++ src/main/java/bowling/domain/Frame.java
@@ -1,6 +1,5 @@
-package bowling;
+package bowling.domain;
 
-import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
