+++ src/test/java/bowling/domain/FrameTest.java
@@ -1,43 +1,39 @@
-package bowling;
+package bowling.domain;
 
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-public class LastFrameTest {
+public class FrameTest {
 
     @Test
-    void last() {
-        assertThat(LastFrame.from()).isNotNull();
+    void from() {
+        assertThat(Frame.from()).isNotNull();
     }
 
     @Test
     void hit_strike() {
-        Frame frame = LastFrame.from();
-        assertThat(frame.hit(10)).isEqualTo("X");
-        assertThat(frame.hit(10)).isEqualTo("X|X");
-        assertThat(frame.hit(10)).isEqualTo("X|X|X");
+        assertThat(Frame.from().hit(10)).isEqualTo("X");
     }
 
     @Test
     void hit_spare() {
-        Frame frame = LastFrame.from();
+        Frame frame = Frame.from();
         assertThat(frame.hit(1)).isEqualTo("1");
         assertThat(frame.hit(9)).isEqualTo("1|/");
-        assertThat(frame.hit(10)).isEqualTo("1|/|X");
     }
 
     @Test
     void hit_miss() {
-        Frame frame = LastFrame.from();
+        Frame frame = Frame.from();
         assertThat(frame.hit(1)).isEqualTo("1");
         assertThat(frame.hit(8)).isEqualTo("1|8");
     }
 
     @Test
     void hit_gutter() {
-        Frame frame = LastFrame.from();
+        Frame frame = Frame.from();
         assertThat(frame.hit(0)).isEqualTo("-");
         assertThat(frame.hit(0)).isEqualTo("-|-");
     }
@@ -45,19 +41,8 @@ public class LastFrameTest {
     @Test
     void hit_exception() {
         assertThatThrownBy(() -> {
-            Frame frame = LastFrame.from();
+            Frame frame = Frame.from();
             frame.hit(0);
-            frame.hit(0);
-            frame.hit(0);
-        }).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void hit_spare_exception() {
-        assertThatThrownBy(() -> {
-            Frame frame = LastFrame.from();
-            frame.hit(1);
-            frame.hit(9);
             frame.hit(0);
             frame.hit(0);
         }).isInstanceOf(IllegalArgumentException.class);
@@ -66,10 +51,8 @@ public class LastFrameTest {
     @Test
     void hit_strike_exception() {
         assertThatThrownBy(() -> {
-            Frame frame = LastFrame.from();
+            Frame frame = Frame.from();
             frame.hit(10);
-            frame.hit(0);
-            frame.hit(0);
             frame.hit(0);
         }).isInstanceOf(IllegalArgumentException.class);
     }
