+++ src/test/java/bowling/domain/LastFrameTest.java
@@ -1,39 +1,43 @@
-package bowling;
+package bowling.domain;
 
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-public class FrameTest {
+public class LastFrameTest {
 
     @Test
-    void from() {
-        assertThat(Frame.from()).isNotNull();
+    void last() {
+        assertThat(LastFrame.from()).isNotNull();
     }
 
     @Test
     void hit_strike() {
-        assertThat(Frame.from().hit(10)).isEqualTo("X");
+        Frame frame = LastFrame.from();
+        assertThat(frame.hit(10)).isEqualTo("X");
+        assertThat(frame.hit(10)).isEqualTo("X|X");
+        assertThat(frame.hit(10)).isEqualTo("X|X|X");
     }
 
     @Test
     void hit_spare() {
-        Frame frame = Frame.from();
+        Frame frame = LastFrame.from();
         assertThat(frame.hit(1)).isEqualTo("1");
         assertThat(frame.hit(9)).isEqualTo("1|/");
+        assertThat(frame.hit(10)).isEqualTo("1|/|X");
     }
 
     @Test
     void hit_miss() {
-        Frame frame = Frame.from();
+        Frame frame = LastFrame.from();
         assertThat(frame.hit(1)).isEqualTo("1");
         assertThat(frame.hit(8)).isEqualTo("1|8");
     }
 
     @Test
     void hit_gutter() {
-        Frame frame = Frame.from();
+        Frame frame = LastFrame.from();
         assertThat(frame.hit(0)).isEqualTo("-");
         assertThat(frame.hit(0)).isEqualTo("-|-");
     }
@@ -41,8 +45,19 @@ public class FrameTest {
     @Test
     void hit_exception() {
         assertThatThrownBy(() -> {
-            Frame frame = Frame.from();
+            Frame frame = LastFrame.from();
             frame.hit(0);
+            frame.hit(0);
+            frame.hit(0);
+        }).isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @Test
+    void hit_spare_exception() {
+        assertThatThrownBy(() -> {
+            Frame frame = LastFrame.from();
+            frame.hit(1);
+            frame.hit(9);
             frame.hit(0);
             frame.hit(0);
         }).isInstanceOf(IllegalArgumentException.class);
@@ -51,8 +66,10 @@ public class FrameTest {
     @Test
     void hit_strike_exception() {
         assertThatThrownBy(() -> {
-            Frame frame = Frame.from();
+            Frame frame = LastFrame.from();
             frame.hit(10);
+            frame.hit(0);
+            frame.hit(0);
             frame.hit(0);
         }).isInstanceOf(IllegalArgumentException.class);
     }
